Fixes way how ContentItemCell is checked of not being reused.
Check of title for match as way to ensure cell wasn't reused is bad because there's no guaranty titles are unique.
UIKit gives method for such check.
If UICollectionView.cellForItem(at:) returns not nil - that is exactly cell which is needed. And returns nil if cell is not visible and there's no reason to update it.